+++ src/html/api-key.html
@@ -242,10 +242,10 @@
       }
 
       .footer a {
-      text-decoration: none;
-      color: inherit;
+        text-decoration: none;
+        color: inherit;
       }
-      
+
       .footer a:hover {
         text-decoration: underline;
       }
@@ -319,12 +319,16 @@
         <img src="../../public/heart.png" alt="heart" height="16" width="16" />
       </span>
       by the
-      <a href="https://github.com/magellan-extension/magellan" target="_blank" rel="noopener noreferrer">
+      <a
+        href="https://github.com/magellan-extension/magellan"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
         Magellan Team
       </a>
     </div>
 
-    <script src="../js/google-ai.js"></script>
-    <script src="../js/api-key.js"></script>
+    <script src="../js/api/google-ai.js"></script>
+    <script src="../js/api/api-key.js"></script>
   </body>
 </html>
