+++ src/html/sidebar.html
@@ -1126,7 +1126,7 @@
         Magellan Team
       </a>
     </div>
-    <script src="../js/google-ai.js" type="module"></script>
-    <script src="../js/sidebar.js" type="module"></script>
+    <script src="../js/api/google-ai.js" type="module"></script>
+    <script src="../js/ui/sidebar.js" type="module"></script>
   </body>
 </html>
