+++ src/js/core/background.js
